Extract stuff from initial add chain
